Full IPC support
The BaseApp has been update to fully support serialised Inter Procedural
Calls. This functionality can be accessed via a new IPC accessor on the
BaseApp.

The dispatch is asynchronous (triggering a callback in the local process
after execution is complete), and is smart enough to detect when an IPC
call can be avoided. In these circumstances execution will be posted to
the local main thread (TID_UI, or TID_RENDERER) for dispatch.

Further work will be required to enhance the IPC data structures with
localised locks to fully support concurrent dispatch with the same process.